feat(frontend): unify message action menus
